[#102]refactor: JwtUtil 작성
