+++ src/main/java/com/sudurukbackback/modulecture/global/security/JwtTokenProvider.java
@@ -2,8 +2,10 @@ package com.sudurukbackback.modulecture.global.security;
 
 import com.sudurukbackback.modulecture.domain.user.entity.User;
 import com.sudurukbackback.modulecture.domain.user.service.AuthService;
+import com.sudurukbackback.modulecture.global.security.util.JwtUtil;
 import io.github.cdimascio.dotenv.Dotenv;
-import io.jsonwebtoken.*;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
 import io.jsonwebtoken.security.Keys;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
@@ -11,16 +13,13 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Component;
 
 import java.nio.charset.StandardCharsets;
 import java.security.Key;
-import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -39,6 +38,7 @@ public class JwtTokenProvider {
 
         String secretKey = getSecretKey();
         this.key = Keys.hmacShaKeyFor(secretKey.getBytes(StandardCharsets.UTF_8));
+        JwtUtil.setKey(key);
     }
 
     // `.env`에서 secretKey 추출
@@ -94,106 +94,12 @@ public class JwtTokenProvider {
         return token;
     }
 
-    /**
-     * JWT 토큰으로부터 인증 객체 생성.
-     *
-     * @param token JWT 토큰.
-     * @return 인증(Authentication) 객체.
-     */
     public Authentication getAuthentication(String token) {
 
-        String username = getUsername(token);
+        String username = JwtUtil.getUsername(token);
+        List<GrantedAuthority> authorities = JwtUtil.getAuthorities(token);
         UserDetails userDetails = authService.loadUserByUsername(username);
 
-        List<GrantedAuthority> authorities = getAuthorities(token);
-
         return new UsernamePasswordAuthenticationToken(userDetails, null, authorities);
-    }
-
-    /**
-     * 토큰에서 사용자 이름(식별자) 추출.
-     *
-     * @param token JWT 토큰.
-     * @return 사용자 이름(식별자).
-     */
-    public String getUsername(String token) {
-
-        String username = parseClaims(token).getSubject();
-        log.debug("Extracted username from JWT: {}", username);
-
-        return username;
-    }
-
-    /**
-     * JWT 토큰에서 권한 정보 추출.
-     *
-     * @param token JWT 토큰.
-     * @return 권한 리스트.
-     */
-    public List<GrantedAuthority> getAuthorities(String token) {
-        // Claims에서 roles 추출
-        Claims claims = parseClaims(token);
-        Object rolesObject = claims.get(KEY_ROLE);
-
-        List<String> roles;
-
-        if (rolesObject instanceof List) {
-            roles = ((List<?>) rolesObject).stream()
-                    .filter(item -> item instanceof String)
-                    .map(item -> (String) item)
-                    .toList();
-        } else if (rolesObject instanceof String) {
-            roles = List.of((String) rolesObject);
-        } else {
-            roles = new ArrayList<>();
-        }
-        log.debug("Extracted roles from JWT: {}", roles);
-
-        return roles.stream()
-                .map(SimpleGrantedAuthority::new)
-                .collect(Collectors.toList());
-    }
-
-    /**
-     * JWT 토큰에서 Claims 추출.
-     *
-     * @param token JWT 토큰.
-     * @return Claims 객체.
-     */
-    private Claims parseClaims(String token) {
-
-        try {
-            return Jwts.parserBuilder()
-                    .setSigningKey(key)
-                    .build()
-                    .parseClaimsJws(token)
-                    .getBody();
-
-        } catch (ExpiredJwtException e) {
-            log.warn("JWT Token expired: {}", e.getClaims().getSubject());
-            return e.getClaims();
-
-        } catch (JwtException e) {
-            log.error("Invalid JWT Token: {}", token);
-            throw new IllegalArgumentException("Invalid JWT token");
-        }
-    }
-
-    /**
-     * JWT 토큰 유효성 검증.
-     *
-     * @param token JWT 토큰.
-     * @return 유효한 토큰이면 true, 그렇지 않으면 false.
-     */
-    public boolean validateToken(String token) {
-
-        try {
-            // Claims를 파싱하여 만료 시간 확인
-            Claims claims = parseClaims(token);
-            return !claims.getExpiration().before(new Date());
-
-        } catch (JwtException | IllegalArgumentException e) {
-            return false;
-        }
     }
 }
